+++ NEWS.md
@@ -1,5 +1,6 @@
 # surveytable (development version)
 
+* `raw` is now a separate argument to `set_opts()`, supported with `output = "screen"`, `"CSV"`, and `"Excel"`. `output = "raw"` is no longer supported.
 * More `set_opts(output = *)` options: `"Word"`, `"Excel"`, `"Excel_v1"`, `"flextable"`.
 * `restructure()`: Restructure tables to make them easier to process programmatically.
 * `set_opts(mode = "nchs")` now sets `drop_na = TRUE`.